// frontend/src/generated/Api.ts
export interface Item {
  /** Name */
  name: string;

  /** Price */
  price: number;

  /** Description */
  description: string;

  /** Image Url */
  image_url: string;

  /**
   * Id
   * @format uuid
   */
  id: string;

  /**
   * Seller Id
   * @format uuid
   */
  seller_id: string;

  /**
   * Created At
   * @format date-time
   */
  created_at: string;

  /**
   * Updated At
   * @format date-time
   */
  updated_at: string;
}